requirements.txt
aiohttp>=3.8.1
python-dotenv>=0.21.0
openai>=0.27.0
pinecone-client>=2.2.0
tqdm>=4.64.0
tiktoken>=0.3.3
beautifulsoup4>=4.11.1
pdfplumber>=0.6.0